Code cleanup: Replace createContentModel with getContentModelCopy
