+++ packages-content-model/roosterjs-content-model-core/lib/corePlugin/ContentModelCopyPastePlugin.ts
@@ -1,13 +1,11 @@
 import { addRangeToSelection } from './utils/addRangeToSelection';
 import { ChangeSource } from '../constants/ChangeSource';
-import { cloneModel } from '../publicApi/model/cloneModel';
 import { deleteEmptyList } from './utils/deleteEmptyList';
 import { deleteSelection } from '../publicApi/selection/deleteSelection';
 import { extractClipboardItems } from '../utils/extractClipboardItems';
 import { getSelectedCells } from '../publicApi/table/getSelectedCells';
 import { iterateSelections } from '../publicApi/selection/iterateSelections';
 import { onCreateCopyEntityNode } from '../override/pasteCopyBlockEntityParser';
-import { transformColor } from '../publicApi/color/transformColor';
 import {
     contentModelToDom,
     createModelToDomContext,
@@ -110,12 +108,7 @@ class ContentModelCopyPastePlugin implements PluginWithState<CopyPastePluginStat
         const selection = this.editor.getDOMSelection();
 
         if (selection && (selection.type != 'range' || !selection.range.collapsed)) {
-            const model = this.editor.createContentModel();
-            const cacheProcessor = this.editor.isDarkMode() ? this.processEntityColor : false;
-
-            const pasteModel = cloneModel(model, {
-                includeCachedElement: cacheProcessor,
-            });
+            const pasteModel = this.editor.getContentModelCopy('disconnected');
 
             if (selection.type === 'table') {
                 iterateSelections(pasteModel, (_, tableContext) => {
@@ -238,25 +231,6 @@ class ContentModelCopyPastePlugin implements PluginWithState<CopyPastePluginStat
 
         return div;
     }
-
-    private processEntityColor = (
-        node: HTMLElement,
-        type: 'general' | 'entity' | 'cache'
-    ): HTMLElement | undefined => {
-        if (type == 'cache' || !this.editor) {
-            return undefined;
-        }
-
-        const result = node.cloneNode(true /*deep*/) as HTMLElement;
-        const colorHandler = this.editor.getColorManager();
-
-        transformColor(result, true /*includeSelf*/, 'darkToLight', colorHandler);
-
-        result.style.color = result.style.color || 'inherit';
-        result.style.backgroundColor = result.style.backgroundColor || 'inherit';
-
-        return result;
-    };
 }
 
 /**
