+++ packages-content-model/roosterjs-content-model-core/lib/corePlugin/EntityPlugin.ts
@@ -175,7 +175,10 @@ class EntityPlugin implements PluginWithState<EntityPluginState> {
     private getChangedEntities(editor: IStandaloneEditor): ChangedEntity[] {
         const result: ChangedEntity[] = [];
 
-        findAllEntities(editor.createContentModel(), result);
+        editor.formatContentModel(model => {
+            findAllEntities(model, result);
+            return false;
+        });
 
         getObjectKeys(this.state.entityMap).forEach(id => {
             const entry = this.state.entityMap[id];
