+++ packages-content-model/roosterjs-content-model-core/lib/editor/StandaloneEditor.ts
@@ -1,7 +1,10 @@
 import { ChangeSource } from '../constants/ChangeSource';
-import { createEmptyModel } from 'roosterjs-content-model-dom';
+import { cloneModel } from '../publicApi/model/cloneModel';
+import { createEmptyModel, tableProcessor } from 'roosterjs-content-model-dom';
 import { createStandaloneEditorCore } from './createStandaloneEditorCore';
+import { reducedModelChildProcessor } from '../override/reducedModelChildProcessor';
 import { transformColor } from '../publicApi/color/transformColor';
+import type { CachedElementHandler } from '../publicApi/model/cloneModel';
 import type {
     ClipboardData,
     ContentModelDocument,
@@ -11,7 +14,6 @@ import type {
     DOMEventRecord,
     DOMHelper,
     DOMSelection,
-    DomToModelOption,
     EditorEnvironment,
     FormatWithContentModelOptions,
     IStandaloneEditor,
@@ -84,15 +86,38 @@ export class StandaloneEditor implements IStandaloneEditor {
 
     /**
      * Create Content Model from DOM tree in this editor
-     * @param option The option to customize the behavior of DOM to Content Model conversion
+     * @param mode What kind of Content Model we want. Currently we support the following values:
+     * - connected: Returns a connect Content Model object. "Connected" means if there is any entity inside editor, the returned Content Model will
+     * contain the same wrapper element for entity. This option should only be used in some special cases. In most cases we should use "disconnected"
+     * to get a fully disconnected Content Model so that any change to the model will not impact editor content.
+     * - disconnected: Returns a disconnected clone of Content Model from editor which you can do any change on it and it won't impact the editor content.
+     * If there is any entity in editor, the returned object will contain cloned copy of entity wrapper element.
+     * If editor is in dark mode, the cloned entity will be converted back to light mode.
+     * - reduced: Returns a reduced Content Model that only contains the model of current selection. If there is already a up-to-date cached model, use it
+     * instead to improve performance. This is mostly used for retrieve current format state.
      */
-    createContentModel(
-        option?: DomToModelOption,
-        selectionOverride?: DOMSelection
-    ): ContentModelDocument {
+    getContentModelCopy(mode: 'connected' | 'disconnected' | 'reduced'): ContentModelDocument {
         const core = this.getCore();
 
-        return core.api.createContentModel(core, option, selectionOverride);
+        switch (mode) {
+            case 'connected':
+                return core.api.createContentModel(core, {
+                    processorOverride: {
+                        table: tableProcessor, // Use the original table processor to create Content Model with real table content but not just an entity
+                    },
+                });
+
+            case 'disconnected':
+                return cloneModel(core.api.createContentModel(core), {
+                    includeCachedElement: this.cloneOptionCallback,
+                });
+            case 'reduced':
+                return core.api.createContentModel(core, {
+                    processorOverride: {
+                        child: reducedModelChildProcessor,
+                    },
+                });
+        }
     }
 
     /**
@@ -396,4 +421,23 @@ export class StandaloneEditor implements IStandaloneEditor {
         }
         return this.core;
     }
+
+    private cloneOptionCallback: CachedElementHandler = (node, type) => {
+        if (type == 'cache') {
+            return undefined;
+        }
+
+        const result = node.cloneNode(true /*deep*/) as HTMLElement;
+
+        if (this.isDarkMode()) {
+            const colorHandler = this.getColorManager();
+
+            transformColor(result, true /*includeSelf*/, 'darkToLight', colorHandler);
+
+            result.style.color = result.style.color || 'inherit';
+            result.style.backgroundColor = result.style.backgroundColor || 'inherit';
+        }
+
+        return result;
+    };
 }
