+++ packages-content-model/roosterjs-content-model-core/test/corePlugin/EntityPluginTest.ts
@@ -3,6 +3,7 @@ import * as transformColor from '../../lib/publicApi/color/transformColor';
 import { createContentModelDocument, createEntity } from '../../../roosterjs-content-model-dom/lib';
 import { createEntityPlugin } from '../../lib/corePlugin/EntityPlugin';
 import {
+    ContentModelDocument,
     DarkColorHandler,
     EntityPluginState,
     IStandaloneEditor,
@@ -12,15 +13,20 @@ import {
 describe('EntityPlugin', () => {
     let editor: IStandaloneEditor;
     let plugin: PluginWithState<EntityPluginState>;
-    let createContentModelSpy: jasmine.Spy;
+    let formatContentModelSpy: jasmine.Spy;
     let triggerPluginEventSpy: jasmine.Spy;
     let isDarkModeSpy: jasmine.Spy;
     let isNodeInEditorSpy: jasmine.Spy;
     let transformColorSpy: jasmine.Spy;
     let mockedDarkColorHandler: DarkColorHandler;
+    let mockedModel: ContentModelDocument;
 
     beforeEach(() => {
-        createContentModelSpy = jasmine.createSpy('createContentModel');
+        formatContentModelSpy = jasmine
+            .createSpy('formatContentModel')
+            .and.callFake((callback: Function) => {
+                callback(mockedModel);
+            });
         triggerPluginEventSpy = jasmine.createSpy('triggerEvent');
         isDarkModeSpy = jasmine.createSpy('isDarkMode');
         isNodeInEditorSpy = jasmine.createSpy('isNodeInEditor');
@@ -28,7 +34,7 @@ describe('EntityPlugin', () => {
         mockedDarkColorHandler = 'COLOR' as any;
 
         editor = {
-            createContentModel: createContentModelSpy,
+            formatContentModel: formatContentModelSpy,
             triggerEvent: triggerPluginEventSpy,
             isDarkMode: isDarkModeSpy,
             isNodeInEditor: isNodeInEditorSpy,
@@ -48,7 +54,7 @@ describe('EntityPlugin', () => {
 
     describe('EditorReady event', () => {
         it('empty doc', () => {
-            createContentModelSpy.and.returnValue(createContentModelDocument());
+            mockedModel = createContentModelDocument();
 
             plugin.onPluginEvent({
                 eventType: 'editorReady',
@@ -68,7 +74,7 @@ describe('EntityPlugin', () => {
 
             doc.blocks.push(entity);
 
-            createContentModelSpy.and.returnValue(doc);
+            mockedModel = doc;
 
             plugin.onPluginEvent({
                 eventType: 'editorReady',
@@ -108,7 +114,7 @@ describe('EntityPlugin', () => {
 
             doc.blocks.push(entity);
 
-            createContentModelSpy.and.returnValue(doc);
+            mockedModel = doc;
             triggerPluginEventSpy.and.returnValue({
                 shouldPersist: true,
             });
@@ -153,7 +159,7 @@ describe('EntityPlugin', () => {
 
             doc.blocks.push(entity);
 
-            createContentModelSpy.and.returnValue(doc);
+            mockedModel = doc;
 
             plugin.onPluginEvent({
                 eventType: 'contentChanged',
@@ -193,7 +199,7 @@ describe('EntityPlugin', () => {
 
             doc.blocks.push(entity);
 
-            createContentModelSpy.and.returnValue(doc);
+            mockedModel = doc;
             isDarkModeSpy.and.returnValue(true);
 
             plugin.onPluginEvent({
@@ -240,7 +246,7 @@ describe('EntityPlugin', () => {
 
             doc.blocks.push(entity);
 
-            createContentModelSpy.and.returnValue(doc);
+            mockedModel = doc;
             const state = plugin.getState();
 
             const wrapper2 = document.createElement('div');
@@ -298,7 +304,7 @@ describe('EntityPlugin', () => {
         it('Do not trigger event for already deleted entity', () => {
             const doc = createContentModelDocument();
 
-            createContentModelSpy.and.returnValue(doc);
+            mockedModel = doc;
             const state = plugin.getState();
 
             const wrapper2 = document.createElement('div');
@@ -332,7 +338,7 @@ describe('EntityPlugin', () => {
 
             doc.blocks.push(entity);
 
-            createContentModelSpy.and.returnValue(doc);
+            mockedModel = doc;
             const state = plugin.getState();
 
             state.entityMap['Entity1'] = {
@@ -519,7 +525,7 @@ describe('EntityPlugin', () => {
                 isReadonly: true,
             });
             doc.blocks.push(entity);
-            createContentModelSpy.and.returnValue(doc);
+            mockedModel = doc;
 
             state.entityMap[id] = {
                 element: wrapper,
