+++ packages-content-model/roosterjs-content-model-core/test/editor/StandaloneEditorTest.ts
@@ -1,8 +1,11 @@
+import * as cloneModel from '../../lib/publicApi/model/cloneModel';
 import * as createEmptyModel from 'roosterjs-content-model-dom/lib/modelApi/creators/createEmptyModel';
 import * as createStandaloneEditorCore from '../../lib/editor/createStandaloneEditorCore';
 import * as transformColor from '../../lib/publicApi/color/transformColor';
 import { ChangeSource } from '../../lib/constants/ChangeSource';
+import { reducedModelChildProcessor } from '../../lib/override/reducedModelChildProcessor';
 import { StandaloneEditor } from '../../lib/editor/StandaloneEditor';
+import { tableProcessor } from 'roosterjs-content-model-dom';
 
 describe('StandaloneEditor', () => {
     let createEditorCoreSpy: jasmine.Spy;
@@ -106,7 +109,7 @@ describe('StandaloneEditor', () => {
         expect(disposeErrorHandlerSpy).toHaveBeenCalledWith(mockedPlugin2, new Error('test'));
     });
 
-    it('createContentModel', () => {
+    it('getContentModelCopy', () => {
         const div = document.createElement('div');
         const mockedModel = 'MODEL' as any;
         const createContentModelSpy = jasmine
@@ -129,25 +132,112 @@ describe('StandaloneEditor', () => {
 
         const editor = new StandaloneEditor(div);
 
-        const model1 = editor.createContentModel();
+        const model1 = editor.getContentModelCopy('connected');
 
         expect(model1).toBe(mockedModel);
-        expect(createContentModelSpy).toHaveBeenCalledWith(mockedCore, undefined, undefined);
+        expect(createContentModelSpy).toHaveBeenCalledWith(mockedCore, {
+            processorOverride: {
+                table: tableProcessor, // Use the original table processor to create Content Model with real table content but not just an entity
+            },
+        });
 
-        const mockedOptions = 'OPTIONS' as any;
-        const selectionOverride = 'SELECTION' as any;
-
-        const model2 = editor.createContentModel(mockedOptions, selectionOverride);
+        const model2 = editor.getContentModelCopy('reduced');
 
         expect(model2).toBe(mockedModel);
-        expect(createContentModelSpy).toHaveBeenCalledWith(
-            mockedCore,
-            mockedOptions,
-            selectionOverride
-        );
+        expect(createContentModelSpy).toHaveBeenCalledWith(mockedCore, {
+            processorOverride: {
+                child: reducedModelChildProcessor,
+            },
+        });
 
         editor.dispose();
-        expect(() => editor.createContentModel()).toThrow();
+        expect(() => editor.getContentModelCopy('connected')).toThrow();
+        expect(() => editor.getContentModelCopy('reduced')).toThrow();
+        expect(resetSpy).toHaveBeenCalledWith();
+    });
+
+    it('getContentModelCopy to return disconnected model', () => {
+        const div = document.createElement('div');
+        const mockedModel = 'MODEL' as any;
+        const mockedClonedModel = 'MODEL2' as any;
+        const createContentModelSpy = jasmine
+            .createSpy('createContentModel')
+            .and.returnValue(mockedModel);
+        const resetSpy = jasmine.createSpy('reset');
+        const mockedCore = {
+            plugins: [],
+            darkColorHandler: {
+                updateKnownColor: updateKnownColorSpy,
+                reset: resetSpy,
+            },
+            lifecycle: {
+                isDarkMode: false,
+            },
+            api: {
+                createContentModel: createContentModelSpy,
+                setContentModel: setContentModelSpy,
+            },
+        } as any;
+
+        createEditorCoreSpy.and.returnValue(mockedCore);
+
+        const editor = new StandaloneEditor(div);
+
+        const cloneModelSpy = spyOn(cloneModel, 'cloneModel').and.returnValue(mockedClonedModel);
+
+        const model = editor.getContentModelCopy('disconnected');
+
+        expect(cloneModelSpy).toHaveBeenCalledWith(mockedModel, {
+            includeCachedElement: jasmine.anything() as any,
+        });
+
+        const transformColorSpy = spyOn(transformColor, 'transformColor');
+        const onClone = cloneModelSpy.calls.argsFor(0)[1]!
+            .includeCachedElement as cloneModel.CachedElementHandler;
+
+        const clonedNode = {
+            style: {
+                backgroundColor: 'red',
+            },
+        } as any;
+        const cloneNodeSpy = jasmine.createSpy('cloneNode').and.returnValue(clonedNode);
+        const mockedNode = {
+            cloneNode: cloneNodeSpy,
+        } as any;
+
+        expect(onClone(mockedNode, 'cache')).toBeUndefined();
+        expect(cloneNodeSpy).not.toHaveBeenCalled();
+
+        // clone entity in light mode
+        expect(onClone(mockedNode, 'entity')).toBe(clonedNode);
+        expect(cloneNodeSpy).toHaveBeenCalledWith(true);
+
+        expect(model).toBe(mockedClonedModel);
+        expect(createContentModelSpy).toHaveBeenCalledWith(mockedCore);
+        expect(transformColorSpy).not.toHaveBeenCalled();
+
+        // Clone in dark mode
+        mockedCore.lifecycle.isDarkMode = true;
+        expect(onClone(mockedNode, 'entity')).toBe(clonedNode);
+        expect(cloneNodeSpy).toHaveBeenCalledWith(true);
+
+        expect(model).toBe(mockedClonedModel);
+        expect(createContentModelSpy).toHaveBeenCalledWith(mockedCore);
+        expect(transformColorSpy).toHaveBeenCalledWith(
+            clonedNode,
+            true,
+            'darkToLight',
+            mockedCore.darkColorHandler
+        );
+        expect(clonedNode).toEqual({
+            style: {
+                color: 'inherit',
+                backgroundColor: 'red',
+            },
+        });
+
+        editor.dispose();
+        expect(() => editor.getContentModelCopy('disconnected')).toThrow();
         expect(resetSpy).toHaveBeenCalledWith();
     });
