+++ packages-content-model/roosterjs-content-model-plugins/test/paste/e2e/cmPasteTest.ts
@@ -1,8 +1,7 @@
 import * as wordFile from '../../../lib/paste/WordDesktop/processPastedContentFromWordDesktop';
-import { ClipboardData, DomToModelOption, IStandaloneEditor } from 'roosterjs-content-model-types';
+import { ClipboardData, IStandaloneEditor } from 'roosterjs-content-model-types';
 import { expectEqual, initEditor } from './testUtils';
 import { itChromeOnly } from 'roosterjs-content-model-dom/test/testUtils';
-import { tableProcessor } from 'roosterjs-content-model-dom';
 
 const ID = 'CM_Paste_E2E';
 
@@ -36,11 +35,7 @@ describe(ID, () => {
 
         editor.pasteFromClipboard(clipboardData);
 
-        const model = editor.createContentModel(<DomToModelOption>{
-            processorOverride: {
-                table: tableProcessor,
-            },
-        });
+        const model = editor.getContentModelCopy('connected');
 
         expectEqual(model, {
             blockGroupType: 'Document',
