+++ packages-content-model/roosterjs-content-model-types/lib/editor/IStandaloneEditor.ts
@@ -9,7 +9,6 @@ import type { Snapshot } from '../parameter/Snapshot';
 import type { ContentModelDocument } from '../group/ContentModelDocument';
 import type { ContentModelSegmentFormat } from '../format/ContentModelSegmentFormat';
 import type { DOMSelection } from '../selection/DOMSelection';
-import type { DomToModelOption } from '../context/DomToModelOption';
 import type { EditorEnvironment } from '../parameter/EditorEnvironment';
 import type {
     ContentModelFormatter,
@@ -25,15 +24,17 @@ import type { TrustedHTMLHandler } from '../parameter/TrustedHTMLHandler';
 export interface IStandaloneEditor {
     /**
      * Create Content Model from DOM tree in this editor
-     * @param rootNode Optional start node. If provided, Content Model will be created from this node (including itself),
-     * otherwise it will create Content Model for the whole content in editor.
-     * @param option The options to customize the behavior of DOM to Content Model conversion
-     * @param selectionOverride When specified, use this selection to override existing selection inside editor
+     * @param mode What kind of Content Model we want. Currently we support the following values:
+     * - connected: Returns a connect Content Model object. "Connected" means if there is any entity inside editor, the returned Content Model will
+     * contain the same wrapper element for entity. This option should only be used in some special cases. In most cases we should use "disconnected"
+     * to get a fully disconnected Content Model so that any change to the model will not impact editor content.
+     * - disconnected: Returns a disconnected clone of Content Model from editor which you can do any change on it and it won't impact the editor content.
+     * If there is any entity in editor, the returned object will contain cloned copy of entity wrapper element.
+     * If editor is in dark mode, the cloned entity will be converted back to light mode.
+     * - reduced: Returns a reduced Content Model that only contains the model of current selection. If there is already a up-to-date cached model, use it
+     * instead to improve performance. This is mostly used for retrieve current format state.
      */
-    createContentModel(
-        option?: DomToModelOption,
-        selectionOverride?: DOMSelection
-    ): ContentModelDocument;
+    getContentModelCopy(mode: 'connected' | 'disconnected' | 'reduced'): ContentModelDocument;
 
     /**
      * Get current running environment, such as if editor is running on Mac
